requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altgraph==0.17.4
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
discord.py==2.4.0
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
pyarmor==9.0.7
pyarmor.cli.core==7.6.3
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
requests==2.32.3
setuptools==75.8.0
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
